update sdk version to 1.0.1

+++ app/src/main/java/com/connect/demo/model/RpcResponse.kt
@@ -4,7 +4,7 @@ package com.connect.demo.model
  * Created by chaichuanfa on 2022/7/27
  */
 data class RpcResponse<T>(
-    val chainId: Int,
+    val chainId: Long,
     val jsonrpc: String,
     val id: String,
     val result: T,
